pyproject.toml
[project]
name = "discord-against-humanity"
version = "0.1.0"
description = "A Cards Against Humanity Discord bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.5",
    "motor>=3.7",
    "html2text>=2024.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.26",
    "pytest-cov>=6.0",
    "ruff>=0.11",
    "mypy>=1.15",
]

[project.scripts]
discord-against-humanity = "discord_against_humanity.bot:main"

[tool.ruff]
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_against_humanity"]
